Control namespace, changed output, method names
Added a TCPGameServer.Control namespace, which contains the parts of the
code that aren't part of the network code or of the model. A
subnamespace is IO, which contains a new Output class which replaces the
Controller.Print output we used to have.

Changed a lot of method names to be capitalized, added a quit command,
changed the way looking works (which breaks the "multiplayer" we had so
far, since you won't detect people walking by, you only look when you
type "l" or when you walk.