Account security increased a bit more
Since account security is so SUPER IMPORTANT in our game, I increased it
a bit more.

Passwords are now rehashed server side before being put in the player
file, so you can't rip the hashed password from the player file and send
it as plain text to get access to an account.